fix structure table order and etc (#43)
- add field product_name
- add field product_image
- add field product_price
- add factory categories, brands and products
- add seeder sliders
- add if logic show image
- replace icon in tab windows
- replace field type integer with decimal

+++ database/migrations/2026_05_07_131630_create_coupons_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->string('code')->unique();
             $table->enum('type', ['percent', 'fixed']);
-            $table->integer('value');
-            $table->integer('cart_value');
+            $table->decimal('value', 12);
+            $table->decimal('cart_value', 12);
             $table->date('expired_date')->nullable();
             $table->timestamps();
         });
